MICCAI2023/multi-view-AE/pyproject.toml
[tool.poetry]
name = "multiviewae"
version = "1.0.0"
description = "A library for running multiview autoencoder models"
authors = ["analawryaguila <ana.lawryaguila@outlook.com>, Alejandra Jayme <alejandra.jayme@icloud.com>"]
readme = "README.md"
packages = [{include = "multiviewae"}]

license = "MIT"
documentation = "https://multiviewae.readthedocs.io/en/latest/"

[tool.setuptools.package-data]
"" = ["*.yaml"]

[tool.poetry.dependencies]
python = "^3.6"
pytest = "^5.4.1"
torch= "^1.10.1"
torchvision = "*"
pytorch-lightning = "1.5.10"
scipy = "*"
pandas = "*"
hydra-core = "*"
schema = "*"
numpy = "<1.24"
matplotlib = "*"
ordereddict = "*"
scikit-learn = "*"
seaborn = "*"
statannotations = "*"
enigmatoolbox = {git = "https://github.com/MICA-MNI/ENIGMA.git", rev = "master"}

[build-system]
requires = ["poetry-core", "setuptools ~= 58.0"]
build-backend = "poetry.core.masonry.api"